+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the Pixel Banner plugin will be documented in this file.
 
+### [v2.19.1] - 2025-01-08
+#### ✨ Added
+- Show example frontmatter values for each field in "Custom Field Names" settings tab
+
+#### 📦 Updated
+- Updated CSS to address padding issue with `Bread Crumbs` and `Typwriter Mode` plugins
+
 ### [v2.19.0] - 2025-01-06
 #### ✨ Added
 - Horizontal image positioning with new X-Position slider